refactor: 상수를 사용되는 클래스로 위치 변경
