+++ src/main/java/domain/Constant.java
@@ -1,12 +1,5 @@
 package domain;
 
 public class Constant {
-    public static int BLACKJACK_MAX_NUMBER = 21;
-    public static int DEFAULT_HAND_NUMBER = 2;
-    public static int ACE_MAX_VALUE = 11;
-    public static int ACE_MIN_VALUE = 1;
-    public static int DEALER_HIT_STAND_BOUNDARY = 16;
-    public static int NAME_MIN_LENGTH = 1;
-    public static int NAME_MAX_LENGTH = 10;
     public static String DELIMITER = ", ";
 }
