+++ src/main/java/domain/Rank.java
@@ -1,11 +1,9 @@
 package domain;
 
-import static domain.Constant.ACE_MAX_VALUE;
-import static domain.Constant.ACE_MIN_VALUE;
-import static domain.Constant.BLACKJACK_MAX_NUMBER;
+import static domain.Result.BLACKJACK_MAX_NUMBER;
 
 public enum Rank {
-    ACE("A", ACE_MIN_VALUE),
+    ACE("A", 1),
     TWO("2", 2),
     THREE("3", 3),
     FOUR("4", 4),
@@ -21,6 +19,8 @@ public enum Rank {
 
     private final String displayValue;
     private final int scoreValue;
+    public static final int ACE_MAX_VALUE = 11;
+    public static final int ACE_MIN_VALUE = 1;
 
     Rank(String displayValue, int scoreValue) {
         this.displayValue = displayValue;
