+++ src/test/java/domain/participant/DealerTest.java
@@ -1,6 +1,6 @@
 package domain.participant;
 
-import static domain.Constant.DEALER_HIT_STAND_BOUNDARY;
+import static domain.BlackjackGame.DEALER_HIT_STAND_BOUNDARY;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 import domain.Rank;
@@ -11,12 +11,12 @@ import org.junit.jupiter.api.Test;
 
 public class DealerTest {
     @Test
-    void 정상적으로_생성되어야_한다(){
+    void 정상적으로_생성되어야_한다() {
         assertDoesNotThrow(() -> new Dealer("딜러"));
     }
 
     @Test
-    void 딜러_카드_중_첫번째_카드만_반환해야_한다(){
+    void 딜러_카드_중_첫번째_카드만_반환해야_한다() {
         Dealer dealer = new Dealer("딜러");
 
         dealer.add(new Card(Suit.CLUB, Rank.K));
@@ -28,7 +28,7 @@ public class DealerTest {
     }
 
     @Test
-    void 딜러의_합계가_기준보다_작으면_참을_반환해야_한다(){
+    void 딜러의_합계가_기준보다_작으면_참을_반환해야_한다() {
         Dealer dealer = new Dealer("딜러");
 
         dealer.add(new Card(Suit.CLUB, Rank.FOUR));
@@ -38,7 +38,7 @@ public class DealerTest {
     }
 
     @Test
-    void 딜러의_합계가_기준보다_크면_거짓을_반환해야_한다(){
+    void 딜러의_합계가_기준보다_크면_거짓을_반환해야_한다() {
         Dealer dealer = new Dealer("딜러");
 
         dealer.add(new Card(Suit.CLUB, Rank.SIX));
